Each block has different metadata for direction, not all the same. Fix and undo some code from previous commit.  How we determined direction was incorrect. Now fixed. Fix respawn packet in protocol.h, respawning still broken. Fix beds: placed and broken correctly Remove useless comments from falling.cpp Re-add chest animation, only one half of double chest works, not sure about animation and multiply people looking in chest. Comment out sign debug message add /time as an alternative to /settime

+++ src/protocol.h
@@ -216,7 +216,7 @@ class Protocol
     static Packet respawn(int world = 0)
     {
       Packet ret;
-      ret << (int8_t)PACKET_RESPAWN << (int8_t)world << (int8_t)1 << (int8_t)0 << (int8_t)128 << (int64_t)0;
+      ret << (int8_t)PACKET_RESPAWN << (int8_t)world << (int8_t)1 << (int8_t)0 << (int16_t)128 << (int64_t)0;
       return ret;
     }
